refactor(world_of_beast): align structure with architecture
