+++ apps/world_of_beast/lib/features/monsters/domain/entities/monster.dart
