+++ apps/world_of_beast/lib/features/monsters/domain/repositories/favorite_monsters_repository.dart
