+++ apps/world_of_beast/lib/features/monsters/domain/repositories/monster_repository.dart
