+++ apps/world_of_beast/lib/features/monsters/domain/usecases/save_favorite_monster_ids_use_case.dart
