+++ apps/world_of_beast/lib/features/monsters/presentation/viewmodels/monster_favorites_controller.dart
